refactor: migrate from telegram_bots to nlp_service
- Updated GitHub Actions workflow to install and run tests for the nlp_service instead of telegram_bots.
- Changed environment variable references from TELEGRAM_BOTS_URL to NLP_SERVICE_URL in postInspector.go.
- Removed the telegram_bots package, including its configuration, Dockerfile, and related files, to streamline the project structure and focus on the nlp_service functionality.
