+++ packages/nlp_service/Dockerfile
@@ -11,10 +11,10 @@ RUN apt-get update && apt-get install -y curl && rm -rf /var/lib/apt/lists/*
 ADD . /app
 
 # Install the required packages using the setup.py in editable mode
-RUN pip install --no-cache-dir -e telegram_bots/
+RUN pip install --no-cache-dir -e nlp_service/
 
 # Copy the .env file into the container
 COPY .env /app/.env
 
 # Run the command to start the main application
-CMD ["python3", "-m", "telegram_bots.main"]
+CMD ["python3", "-m", "nlp_service.main"]
