+++ packages/nlp_service/FreiFahren_BE_NLP/data/synonyms.json
